refactor(lib): extract lib/auth_rate_limit.php — login/IP rate limiting + lockout (ADR-004, sub of #907)
Moves the login/IP rate-limiting and account-lockout functions out of lib.php
into lib/auth_rate_limit.php. ADR-003 conversion applied. Plain extraction —
Bug Y #1122 was already fixed in v3.27.2 (ipam_sudo_would_strand_user_after_
disable + MfaDisableLockoutGuardTest), so no Bug Y work is in scope; #1122 can
be closed as already-resolved. reCAPTCHA functions stay behind for Task 6.4.
