+++ Simple-PHP-IPAM/lib.php
@@ -11,6 +11,7 @@ require_once __DIR__ . '/lib/settings.php';
 require_once __DIR__ . '/lib/user_preferences.php';
 require_once __DIR__ . '/lib/auth.php';     // core session + CSRF + login (ADR-004 Task 6.1)
 require_once __DIR__ . '/lib/auth_password.php'; // password policy + reset tokens (ADR-004 Task 6.2)
+require_once __DIR__ . '/lib/auth_rate_limit.php'; // login/IP rate limiting + lockout (ADR-004 Task 6.3)
 require_once __DIR__ . '/lib/BackupClientInterface.php';
 require_once __DIR__ . '/lib/S3Client.php';
 require_once __DIR__ . '/lib/SftpClient.php';
@@ -132,246 +133,18 @@ function ipam_format_datetime(string|int|null $utc, ?string $fmt = null, ?int $u
 /* current_user(), require_role(), require_write_access(), and login_user()  */
 /* moved to lib/auth.php in v3.30.0 (ADR-004 Phase 6 Task 6.1, #907).        */
 
-/* ---------------- Login rate limiting ---------------- */
-
-/**
- * Generic per-action, per-IP rate limiter (#882). Counts rows in
- * login_attempts whose action matches and whose attempted_at is within the
- * sliding window. The legacy login_*-named helpers below remain as
- * action='login' wrappers so existing callers and tests are unaffected.
- */
-function auth_rate_limited(PDO $db, string $action, string $ip, int $maxAttempts, int $windowSeconds): bool
-{
-    $cutoff = date('Y-m-d H:i:s', time() - $windowSeconds);
-    $st = $db->prepare(
-        "SELECT COUNT(*) AS c FROM login_attempts
-          WHERE action = :a AND ip = :ip AND attempted_at >= :cutoff"
-    );
-    $st->execute([':a' => $action, ':ip' => $ip, ':cutoff' => $cutoff]);
-    /** @var array<string, mixed>|false $countRow */
-    $countRow = $st->fetch();
-    return (is_array($countRow) ? to_int($countRow['c']) : 0) >= $maxAttempts;
-}
-
-function record_auth_failure(PDO $db, string $action, string $ip, string $username = ''): void
-{
-    $db->prepare(
-        "INSERT INTO login_attempts (ip, username, action) VALUES (:ip, :username, :a)"
-    )->execute([
-        ':ip'       => $ip,
-        ':username' => $username !== '' ? $username : null,
-        ':a'        => $action,
-    ]);
-}
-
-function clear_auth_failures(PDO $db, string $action, string $ip): void
-{
-    $db->prepare("DELETE FROM login_attempts WHERE action = :a AND ip = :ip")
-       ->execute([':a' => $action, ':ip' => $ip]);
-}
-
-function login_rate_limited(PDO $db, string $ip, int $maxAttempts, int $windowSeconds): bool
-{
-    return auth_rate_limited($db, 'login', $ip, $maxAttempts, $windowSeconds);
-}
-
-/**
- * Real lockout-expiry timestamp for a rolling-window IP rate limit.
- *
- * The window is "max N failures in the past $windowSeconds." The unlock
- * moment is when the OLDEST currently-counted failure ages out — at that
- * point the window has only N-1 failures and a fresh attempt is allowed.
- * Returns time() + $windowSeconds as a sane fallback when no attempts
- * are recorded (caller is asking for a future window without any
- * existing data — shouldn't happen for actively-rate-limited IPs).
- *
- * Introduced for CR PR #1141: pre-fix, login.php sent
- * `time() + $lockoutSeconds` to ipam_audit_ip_rate_limited(), which
- * overstated the wait under steady traffic and could keep the dampener
- * suppressing past the real unlock point.
- */
-function auth_rate_limit_unlock_at(PDO $db, string $action, string $ip, int $maxAttempts, int $windowSeconds): int
-{
-    $cutoff = date('Y-m-d H:i:s', time() - $windowSeconds);
-    // CR PR #1141 round 2: locate the THRESHOLD-CROSSING failure (the
-    // Nth-from-newest in the window). Pre-fix used MIN() of all
-    // in-window failures, which is only correct when exactly N failures
-    // exist. With more than N, expiring the absolute oldest still
-    // leaves the IP blocked — `unlock_at` was reported too early and
-    // the dampener could roll over before the real unlock point.
-    $countSt = $db->prepare(
-        "SELECT COUNT(*) FROM login_attempts
-          WHERE action = :a AND ip = :ip AND attempted_at >= :cutoff"
-    );
-    $countSt->execute([':a' => $action, ':ip' => $ip, ':cutoff' => $cutoff]);
-    $count = to_int($countSt->fetchColumn());
-    if ($count === 0) {
-        return time() + $windowSeconds;
-    }
-    // OFFSET = count - maxAttempts; e.g. with N=5 max and 7 in-window
-    // failures, the threshold-crossing row is at OFFSET 2 (skipping the
-    // 2 oldest that are about to age out before the lockout actually
-    // lifts). With exactly N in-window failures, OFFSET = 0 == oldest.
-    $offset = max(0, $count - $maxAttempts);
-    $st = $db->prepare(
-        "SELECT attempted_at FROM login_attempts
-          WHERE action = :a AND ip = :ip AND attempted_at >= :cutoff
-          ORDER BY attempted_at ASC
-          LIMIT 1 OFFSET :offset"
-    );
-    $st->bindValue(':a',      $action,  PDO::PARAM_STR);
-    $st->bindValue(':ip',     $ip,      PDO::PARAM_STR);
-    $st->bindValue(':cutoff', $cutoff,  PDO::PARAM_STR);
-    $st->bindValue(':offset', $offset,  PDO::PARAM_INT);
-    $st->execute();
-    $oldest = to_str($st->fetchColumn());
-    if ($oldest === '') {
-        return time() + $windowSeconds;
-    }
-    $oldestTs = strtotime($oldest);
-    if ($oldestTs === false) {
-        return time() + $windowSeconds;
-    }
-    return $oldestTs + $windowSeconds;
-}
-
-/**
- * Emit a once-per-window 'auth.ip_rate_limited' audit row (#1134, v3.27.3;
- * made atomic in v3.28.0 #1143).
- *
- * Pre-fix (#1134): login.php audited 'auth.login_blocked' on every refused
- * attempt within the rate-limit window — flooding the log with up to one
- * row per attempted login over the lockout window. Operators filtering
- * audit_log by username also missed these rows entirely (entity_id NULL,
- * username column NULL).
- *
- * #1134 contract: emit a single row per (action, ip) lockout window with
- * details = "action=$action attempts=$attempts unlock_at=$ISO8601 ip=$ip"
- * so the operator can grep it on either the audit_log.ip column or the
- * details substring; subsequent attempts inside the same window do not
- * re-emit.
- *
- * #1143: the original dampener enforced "once per window" by SELECTing the
- * most recent prior 'auth.ip_rate_limited' row out of audit_log and only
- * INSERTing if no active window was found — a read-then-insert TOCTOU that
- * let two concurrent brute-force requests both miss the prior row and both
- * emit. The window is now tracked in the `rate_limit_dampener` table
- * (PRIMARY KEY (action, ip), unlock_at = Unix epoch). The caller claims
- * the window with an UPDATE that only fires on an expired row, falling
- * back to an INSERT when no row exists at all; exactly one concurrent
- * caller wins the INSERT (PK collision), so at most one audit row is
- * emitted per window regardless of concurrency.
- */
-function ipam_audit_ip_rate_limited(PDO $db, string $action, string $ip, int $attempts, int $unlockAt): void
-{
-    $now = time();
-
-    // Claim this (action, ip) lockout window atomically. A row is "active"
-    // while unlock_at > now; we may claim the window only when no active
-    // row exists. Step 1: UPDATE an expired row to the new window (fires
-    // exactly when a row exists AND it has expired). Step 2 (only if the
-    // UPDATE matched nothing): INSERT a fresh row — succeeds when no row
-    // exists, raises a UNIQUE violation when an active row already holds
-    // the window (or a concurrent caller just inserted one), which we
-    // treat as "already dampened".
-    $upd = $db->prepare(
-        "UPDATE rate_limit_dampener SET unlock_at = :u
-          WHERE action = :a AND ip = :ip AND unlock_at <= :now"
-    );
-    $upd->execute([':u' => $unlockAt, ':a' => $action, ':ip' => $ip, ':now' => $now]);
-    $claimed = $upd->rowCount() > 0;
-
-    if (!$claimed) {
-        try {
-            $ins = $db->prepare(
-                "INSERT INTO rate_limit_dampener (action, ip, unlock_at) VALUES (:a, :ip, :u)"
-            );
-            $ins->execute([':a' => $action, ':ip' => $ip, ':u' => $unlockAt]);
-            $claimed = true;
-        } catch (\PDOException $e) {
-            $sqlstate = (string) ($e->errorInfo[0] ?? '');
-            $msg = $e->getMessage();
-            $isUniqueViolation = $sqlstate === '23000' || $sqlstate === '23505'
-                || stripos($msg, 'unique') !== false
-                || stripos($msg, 'duplicate') !== false;
-            if (!$isUniqueViolation) {
-                throw $e;
-            }
-            // Active row already holds the window — dampened.
-        }
-    }
-
-    if (!$claimed) {
-        return;
-    }
-
-    $unlockIso = gmdate('Y-m-d\TH:i:s\Z', $unlockAt);
-    $emitted = audit($db, 'auth.ip_rate_limited', 'auth', null,
-          "action=$action attempts=$attempts unlock_at=$unlockIso ip=$ip");
-    if (!$emitted) {
-        // audit() returns false (not an exception) on write failure. We just
-        // claimed this (action, ip) window via the UPDATE/INSERT above, so the
-        // claim is the only thing standing between subsequent attempts and a
-        // (still missing) audit row. Roll the claim back so a later attempt in
-        // the same window can re-attempt the emit.
-        $db->prepare("DELETE FROM rate_limit_dampener WHERE action = :a AND ip = :ip")
-           ->execute([':a' => $action, ':ip' => $ip]);
-    }
-}
-
-/**
- * Housekeeping: drop rate_limit_dampener rows whose lockout window expired
- * more than $graceSeconds ago. The table is bounded by the number of
- * distinct (action, ip) pairs that have ever been rate-limited; a sustained
- * brute-force from many sources can grow it, so prune it during the lazy
- * housekeeping sweep alongside purge_old_login_attempts().
- */
-function prune_rate_limit_dampener(PDO $db, int $graceSeconds = 86400): void
-{
-    $cutoff = time() - max(0, $graceSeconds);
-    $db->prepare("DELETE FROM rate_limit_dampener WHERE unlock_at < :cutoff")
-       ->execute([':cutoff' => $cutoff]);
-}
-
-function record_login_failure(PDO $db, string $ip, string $username = ''): void
-{
-    record_auth_failure($db, 'login', $ip, $username);
-}
-
-function clear_login_failures(PDO $db, string $ip): void
-{
-    clear_auth_failures($db, 'login', $ip);
-}
-
-function account_locked_out(PDO $db, string $username, int $maxAttempts, int $windowSeconds): bool
-{
-    $cutoff = date('Y-m-d H:i:s', time() - $windowSeconds);
-    $st = $db->prepare(
-        "SELECT COUNT(*) AS c FROM login_attempts WHERE username = :u AND attempted_at >= :cutoff"
-    );
-    $st->execute([':u' => $username, ':cutoff' => $cutoff]);
-    /** @var array<string, mixed>|false $row */
-    $row = $st->fetch();
-    return (is_array($row) ? to_int($row['c']) : 0) >= $maxAttempts;
-}
-
-function clear_account_lockout(PDO $db, string $username): void
-{
-    $db->prepare("DELETE FROM login_attempts WHERE username = :u")
-       ->execute([':u' => $username]);
-}
+/* Login/IP rate-limiting + account-lockout helpers moved to              */
+/* lib/auth_rate_limit.php in v3.30.0 (ADR-004 Phase 6 Task 6.3, #907):   */
+/* auth_rate_limited(), record_auth_failure(), clear_auth_failures(),     */
+/* login_rate_limited(), auth_rate_limit_unlock_at(),                     */
+/* ipam_audit_ip_rate_limited(), prune_rate_limit_dampener(),             */
+/* record_login_failure(), clear_login_failures(), account_locked_out(),  */
+/* clear_account_lockout(), purge_old_login_attempts().                   */
 
 /** @param IpamConfig $config */
 function recovery_mode_enabled(array $config): bool
 {
     return (bool)($config['recovery_mode'] ?? false);
-}
-
-function purge_old_login_attempts(PDO $db, int $windowSeconds): void
-{
-    $cutoff = date('Y-m-d H:i:s', time() - $windowSeconds);
-    $db->prepare("DELETE FROM login_attempts WHERE attempted_at < :cutoff")
-       ->execute([':cutoff' => $cutoff]);
 }
 
 /* logout_user() and client_ip() moved to lib/auth.php in v3.30.0           */
@@ -7784,84 +7557,9 @@ function ipam_totp_verify_backup_code(PDO $db, int $userId, string $code): bool 
 
 // ============================================================
 // API per-key rate limiting (v3.6.0, #419)
+//   ipam_api_key_rate_limit_check() moved to lib/auth_rate_limit.php
+//   in v3.30.0 (ADR-004 Phase 6 Task 6.3, #907).
 // ============================================================
-
-/**
- * Returns 0 when the request is allowed, or the number of seconds until the
- * sliding window unblocks when it is denied.
- */
-function ipam_api_key_rate_limit_check(PDO $db, string $bucketKey, int $windowSec, int $max): int {
-    if ($windowSec < 1) {
-        throw new \InvalidArgumentException('windowSec must be >= 1');
-    }
-    if ($max < 1) {
-        throw new \InvalidArgumentException('max must be >= 1');
-    }
-    $now              = time();
-    $windowStartEpoch = (int)($now / $windowSec) * $windowSec;
-    $prevWindowEpoch  = $windowStartEpoch - $windowSec;
-    $windowStart      = date('Y-m-d H:i:s', $windowStartEpoch);
-    $prevWindowStart  = date('Y-m-d H:i:s', $prevWindowEpoch);
-    $cutoff           = date('Y-m-d H:i:s', $prevWindowEpoch - $windowSec);
-
-    // Prune buckets older than 2 windows ago
-    $db->prepare("DELETE FROM rate_limit_buckets WHERE window_start < :cutoff")
-       ->execute([':cutoff' => $cutoff]);
-
-    // Increment current window
-    $driver = $db->getAttribute(PDO::ATTR_DRIVER_NAME);
-    if ($driver === 'mysql') {
-        $db->prepare(
-            "INSERT INTO rate_limit_buckets (bucket_key, window_start, count) VALUES (:k, :w, 1)
-             ON DUPLICATE KEY UPDATE count = count + 1"
-        )->execute([':k' => $bucketKey, ':w' => $windowStart]);
-    } else {
-        $db->prepare(
-            "INSERT INTO rate_limit_buckets (bucket_key, window_start, count) VALUES (:k, :w, 1)
-             ON CONFLICT(bucket_key, window_start) DO UPDATE SET count = rate_limit_buckets.count + 1"
-        )->execute([':k' => $bucketKey, ':w' => $windowStart]);
-    }
-
-    // Sliding-window approximation: weight previous bucket by fraction remaining in current window
-    $stmt = $db->prepare(
-        "SELECT count FROM rate_limit_buckets WHERE bucket_key = :k AND window_start = :w"
-    );
-    $stmt->execute([':k' => $bucketKey, ':w' => $windowStart]);
-    $currentCount = (int)($stmt->fetchColumn() ?: 0);
-
-    $stmt->execute([':k' => $bucketKey, ':w' => $prevWindowStart]);
-    $prevCount = (int)($stmt->fetchColumn() ?: 0);
-
-    $elapsed  = $now - $windowStartEpoch;
-    $weight   = ($windowSec - $elapsed) / $windowSec;
-    $weighted = $currentCount + ($prevCount * $weight);
-
-    if ($weighted <= $max) {
-        return 0; // allowed
-    }
-
-    // Calculate when the weighted count will drop to <= max (assuming no new requests).
-    //
-    // Case A: currentCount < max — overflow is driven by prevCount weight.
-    //   Solve: currentCount + prevCount*(windowSec - elapsed - t)/windowSec = max
-    //   => t = (windowSec - elapsed) - (max - currentCount)*windowSec/prevCount
-    if ($currentCount < $max && $prevCount > 0) {
-        $tWithin = ($windowSec - $elapsed) - ((float)($max - $currentCount) * $windowSec / $prevCount);
-        if ($tWithin > 0) {
-            return max(1, (int)ceil($tWithin));
-        }
-    }
-
-    // Case B: currentCount >= max — need to survive into the next window.
-    //   In next window: weighted ≈ currentCount*(windowSec - t')/windowSec
-    //   Solve for t' when weighted = max, then add remaining time in current window.
-    $remainInWindow = $windowStartEpoch + $windowSec - $now;
-    if ($currentCount > 0) {
-        $intoNext = (int)ceil($windowSec * (1.0 - (float)$max / $currentCount));
-        return max(1, $remainInWindow + max(0, $intoNext));
-    }
-    return max(1, $remainInWindow + 1);
-}
 
 // ============================================================
 // Session absolute lifetime moved to lib/auth.php in v3.30.0
@@ -7869,7 +7567,7 @@ function ipam_api_key_rate_limit_check(PDO $db, string $bucketKey, int $windowSe
 // ============================================================
 
 // ============================================================
-// Persistent account lockout helpers (v3.6.0, #421)
+// Persistent account lockout helpers (v3.6.0, #421) (ipam_clear_persistent_lockout moved to lib/auth_rate_limit.php — Task 6.3)
 // ============================================================
 
 function ipam_is_persistently_locked(PDO $db, int $uid): bool {
@@ -7898,11 +7596,8 @@ function ipam_record_2fa_failure(PDO $db, int $uid, array $config): void {
     }
 }
 
-function ipam_clear_persistent_lockout(PDO $db, int $uid): void {
-    $db->prepare(
-        "UPDATE users SET failed_auth_count = 0, locked_until = NULL, lock_reason = NULL WHERE id = :id"
-    )->execute([':id' => $uid]);
-}
+/* ipam_clear_persistent_lockout() moved to lib/auth_rate_limit.php in     */
+/* v3.30.0 (ADR-004 Phase 6 Task 6.3, #907).                              */
 
 // ============================================================
 // Email OTP helpers (#684)
